FIX мне лень писать

+++ Complete.php
@@ -1,39 +1,40 @@
 <?php
 /**
  * Created by PhpStorm.
- * User: fulle
+ * User: Sariel Aki ~desu~
  * Date: 14.05.2018
  * Time: 22:03
+ *
  */
 
 class Complete
 {
     public static function client(Base $f3){
-        $values_from_client = [];
-        if($f3->get('SESSION.client_order') != null) {
-            array_push($values_from_client, $f3->get('SESSION.client_order'), $f3->get('SESSION.order_value'), $f3->get('SESSION.order_id'), date('y-m-d'));
-            $f3->get('DB')->exec('INSERT INTO executed_orders(client_number, value, order_id, date) VALUE(?, ?, ?, ?)',
-                array_values($values_from_client)
-            );
-            $f3->get('DB')->exec("UPDATE orders SET client_complete_check = 'true' WHERE client_number = ?",
-                $f3->get('SESSION.client_order')
+        if($f3->get('SESSION.order_id') != null) {
+            $f3->get('DB')->exec("UPDATE orders SET client_complete_check = 'true' WHERE id = ?",
+                $f3->get('SESSION.order_id')
+
             );
         }
     }
     public static function driver(Base $f3){
-        $values_from_driver = [];
-        if($f3->get('SESSION.order') != null){
-            array_push($values_from_driver, $f3->get('SESSION.driver_phone'), $f3->get('SESSION.order'));
-            $f3->get('DB')->exec('UPDATE executed_orders SET driver_phone = ? WHERE order_id = ?',
-                array_values($values_from_driver)
-            );
+        if($f3->get('SESSION.order_id') != null){
             $f3->get('DB')->exec("UPDATE orders SET driver_complete_check = 'true' WHERE id = ?",
-                $f3->get('SESSION.order')
+                $f3->get('SESSION.order_id')
             );
         }
     }
 
     static function afterroute(Base $f3) {
-        $f3->get('DB')->exec("DELETE FROM orders WHERE client_complete_check = 'true' AND driver_complete_check = 'true'");
+        $result = $f3->get('DB')->exec("SELECT * FROM orders WHERE client_complete_check = 'true' AND driver_complete_check = 'true'");
+        $json_res =  json_encode($result);
+        $json_code = json_decode($json_res, true);
+        $completed_order_data = array($json_code[0]['id'], $json_code[0]['date'], $json_code[0]['value'], $json_code[0]['driver_phone'], $json_code[0]['client_number']);
+        if(count($result) != 0) {
+            $f3->get('DB')->exec('INSERT INTO executed_orders(order_id, date, value, driver_phone, client_number) VALUE(?, ?, ?, ?, ?)',
+                $completed_order_data
+            );
+            $f3->get('DB')->exec("DELETE FROM orders WHERE client_complete_check = 'true' AND driver_complete_check = 'true'");
+        }
     }
 }
